requirements.txt
# FastAPI Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database Clients
pymilvus==2.3.5
elasticsearch==8.11.1

# AI/ML
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
pillow==10.2.0

# Translation Services
google-generativeai>=0.3.0
deep-translator==1.11.4

# Video Processing
opencv-python==4.9.0.80

# Utilities
numpy==1.26.3
requests==2.31.0

# Fix marshmallow compatibility issue
marshmallow>=3.20.0,<4.0.0
apispec>=6.0.0

# Crawler dependencies
selenium>=4.15.0
selenium-stealth>=1.0.6
webdriver-manager>=4.0.0

# Optional: For better performance
# aiofiles==23.2.1
